refactor(core): consolidate config version consts
Single definition site for the remote config file version, the app-config
schema version (folding settings' latestConfigVersion and config-upgrade's
targetConfigVersion, whose equality is intentional) and the connection-config
versions, in core/constants/config-versions.const.ts (PCS-07). Also corrects
the wrong 12->13 version-space comment in ConfigurationUpgradeService and the
storage patchConfig debug log that claimed a suppression the code never had.

Refs #17
